Update user-related classes in authentication domain
Several updates have been made to the User-related classes in the authentication domain. These updates involve the addition of summary comments for clarity, the adjustment of class constructors and their parameters for better organization, as well as the introduction of new methods to generate random usernames and manage user information. Furthermore, clean-up of code and detailed explanations were added where necessary to improve readability and maintainability.